fix: 1차 구현 분리
